Add cmd tool to send a single TCP probe to a port
This sends a SYN packet to the host name or IPv4 address in os.Args[1]
to the port at os.Args[2].

If it sees a TCP handshake (SYN-ACK) or reset (RST) packet back from
that address on to the port from which it sent then it will print
a little info about it.

But mostly it should just silently read packets that it discards because
there's no reason yet for matching packets.