Set timeout for reader
I changed pcap.BlockForever in the stream to have a timeout. Since the
read is in a loop that only stops on a <-done signal this just has the
effect of throwing us back through the loop where we can check that done
channel again.

The pipeline has structure now, so I'm going to address race conditions
and then move on to concurrent probes.