Add concurrent probe sender
We now have a concurrent prober that starts a goroutine to listen on a
ports (int) channel and is preconfigured with the packets that it will
send. It will tie up when we close the ports channel. At present it's
quite detached from the scanner itself and I'll be refactoring it pretty
quickly I'm sure.