Clean up main - working functional reference
Just tuning the code up with a little refactor.
This can now serve as the reference implementation for putting
the real architecture in place that will provide an interface
to a couple of front-ends. But the real goal of mulling the architecture
was to build my confidence that this would be adequate to build
applications that deal with large scale infrastructures.